transpose data in predict appropriately
When doing things with resampling, the predicted data will come in as a
`data.frame`, but it will be in a wide format (more conducive to formula
stuff) rather than a tall format (more typically the input format).
To get around this, an argument "format" was added with the default
being set to "wide". Users can supply "tall" if using predict more interactively